# openapi.json
{"openapi":"3.1.0","info":{"title":"Magicpin AI Challenge Bot","version":"0.1.0"},"paths":{"/v1/healthz":{"get":{"summary":"Healthz","operationId":"healthz_v1_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/v1/metadata":{"get":{"summary":"Metadata","operationId":"metadata_v1_metadata_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MetadataResponse"}}}}}}},"/v1/context":{"post":{"summary":"Push Context","operationId":"push_context_v1_context_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContextRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContextResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/tick":{"post":{"summary":"Tick","operationId":"tick_v1_tick_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TickRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TickResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/reply":{"post":{"summary":"Reply","operationId":"reply_v1_reply_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplyRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ContextRequest":{"properties":{"scope":{"type":"string","enum":["category","merchant","customer","trigger"],"title":"Scope"},"context_id":{"type":"string","minLength":1,"title":"Context Id"},"version":{"type":"integer","minimum":1.0,"title":"Version"},"payload":{"additionalProperties":true,"type":"object","title":"Payload"},"delivered_at":{"type":"string","format":"date-time","title":"Delivered At"}},"type":"object","required":["scope","context_id","version","payload","delivered_at"],"title":"ContextRequest"},"ContextResponse":{"properties":{"accepted":{"type":"boolean","title":"Accepted"},"ack_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ack Id"},"stored_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Stored At"},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason"},"current_version":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Current Version"},"details":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Details"}},"type":"object","required":["accepted"],"title":"ContextResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","const":"ok","title":"Status"},"uptime_seconds":{"type":"integer","minimum":0.0,"title":"Uptime Seconds"},"contexts_loaded":{"additionalProperties":{"type":"integer"},"propertyNames":{"enum":["category","merchant","customer","trigger"]},"type":"object","title":"Contexts Loaded"}},"type":"object","required":["status","uptime_seconds","contexts_loaded"],"title":"HealthResponse"},"MetadataResponse":{"properties":{"team_name":{"type":"string","title":"Team Name"},"team_members":{"items":{"type":"string"},"type":"array","title":"Team Members"},"model":{"type":"string","title":"Model"},"approach":{"type":"string","title":"Approach"},"contact_email":{"type":"string","title":"Contact Email"},"version":{"type":"string","title":"Version"},"submitted_at":{"type":"string","title":"Submitted At"}},"type":"object","required":["team_name","team_members","model","approach","contact_email","version","submitted_at"],"title":"MetadataResponse"},"ReplyRequest":{"properties":{"conversation_id":{"type":"string","minLength":1,"title":"Conversation Id"},"merchant_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Merchant Id"},"customer_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Customer Id"},"from_role":{"type":"string","enum":["merchant","customer"],"title":"From Role"},"message":{"type":"string","title":"Message"},"received_at":{"type":"string","format":"date-time","title":"Received At"},"turn_number":{"type":"integer","minimum":1.0,"title":"Turn Number"}},"type":"object","required":["conversation_id","from_role","message","received_at","turn_number"],"title":"ReplyRequest"},"ReplyResponse":{"properties":{"action":{"type":"string","enum":["send","wait","end"],"title":"Action"},"body":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Body"},"cta":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cta"},"wait_seconds":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Wait Seconds"},"rationale":{"type":"string","minLength":1,"title":"Rationale"}},"type":"object","required":["action","rationale"],"title":"ReplyResponse"},"TickAction":{"properties":{"conversation_id":{"type":"string","title":"Conversation Id"},"merchant_id":{"type":"string","title":"Merchant Id"},"customer_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Customer Id"},"send_as":{"type":"string","enum":["vera","merchant_on_behalf"],"title":"Send As"},"trigger_id":{"type":"string","title":"Trigger Id"},"template_name":{"type":"string","title":"Template Name"},"template_params":{"items":{"type":"string"},"type":"array","title":"Template Params"},"body":{"type":"string","minLength":1,"title":"Body"},"cta":{"type":"string","title":"Cta"},"suppression_key":{"type":"string","title":"Suppression Key"},"rationale":{"type":"string","minLength":1,"title":"Rationale"}},"type":"object","required":["conversation_id","merchant_id","send_as","trigger_id","template_name","template_params","body","cta","suppression_key","rationale"],"title":"TickAction"},"TickRequest":{"properties":{"now":{"type":"string","format":"date-time","title":"Now"},"available_triggers":{"items":{"type":"string"},"type":"array","title":"Available Triggers"}},"type":"object","required":["now"],"title":"TickRequest"},"TickResponse":{"properties":{"actions":{"items":{"$ref":"#/components/schemas/TickAction"},"type":"array","title":"Actions"}},"type":"object","title":"TickResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}